requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes --output-file requirements-simple.txt
dnspython==2.7.0
    # via pymongo
dotenv==0.9.9
    # via metabase-poc
numpy==2.3.2
    # via pandas
pandas==2.3.2
    # via metabase-poc
psycopg2-binary==2.9.10
    # via metabase-poc
pymongo==4.14.1
    # via metabase-poc
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via dotenv
pytz==2025.2
    # via pandas
six==1.17.0
    # via python-dateutil
tzdata==2025.2
    # via pandas
